# Use MlasComputeSoftmax in traditional ML ops